filters, styles

+++ angular-todo/src/app/todo/todo.component.ts
@@ -12,39 +12,73 @@ import { Todo } from './todo';
 export class TodoComponent implements OnInit {
   todos: Todo[] = [];
   newTodo: any = { name: '' };
-  pageTitle: string;
+  pageTitle: string = 'Mes todos';
+  enableDeleteMulti: boolean = false;
+  loading: boolean = true;
   constructor(private api: ApiService) { }
 
   ngOnInit() {
     this.api.getAll('todos').subscribe(
         (todos) => {
           this.todos = todos;
+          this.loading = false;
         }
       );
   }
 
+  /**
+  * [Add element to todos]
+  */
   addElement() {
+    this.enableDeleteMulti = false;
     const lastElement = this.todos.length - 1;
-    if (this.todos && !this.todos[lastElement].name) return false;
-    this.todos.unshift(new Todo());
+    if (this.todos.length > 0 && !this.todos[lastElement].name) return false;
+    this.todos.unshift(new Todo({ isDone: false }));
   }
 
-  // addElement() {
-  //   if (this.newTodo && !this.newTodo.name) return false;
-  //   this.todos.push(Object.assign({}, this.newTodo));
-  //   this.newTodo.name = '';
-  // }
+  /**
+  * [Detect change on element]
+  */
+  detectChange(element, i) {
+    this.enableDeleteMulti = false;
+    this.todos[i].hasChanged = true;
+  }
 
+  /**
+  * [Mark as done or not and save]
+  */
+  toggleSelected(row, i) {
+    this.todos[i].selected = !row.selected;
+  }
+
+  /**
+  * [Mark as done or not and save]
+  */
+  toggleDone(row, i) {
+    this.enableDeleteMulti = false;
+    this.todos[i].isDone = !row.isDone;
+    this.api.updateOne('todos', row._id, row).subscribe((res) => this.todos[i] );
+  }
+
+  /**
+  * [Save one Element]
+  */
   saveElement(row, index) {
+    this.enableDeleteMulti = false;
     if (!row._id) {
       this.api.createOne('todos', row).subscribe((res) =>
         this.todos[index] = new Todo(res)
       );
       return true;
     }
-    this.api.updateOne('todos', row._id, row).subscribe((res) => row = res);
+    this.api.updateOne('todos', row._id, row).subscribe((res) => this.todos[index].hasChanged = false );
   }
+
+  /**
+  * [delete one Element]
+  */
   deleteElement(id, index) {
+    this.enableDeleteMulti = false;
     this.api.deleteOne('todos', id).subscribe(
         (res) => {
           this.todos.splice(index, 1);
@@ -52,7 +86,14 @@ export class TodoComponent implements OnInit {
       );
   }
 
+  /**
+  * [delete Multiple element if they are selected. Also enable selection]
+  */
   deleteMulti() {
+    if (!this.enableDeleteMulti) {
+      this.enableDeleteMulti = true;
+      return true
+    }
     const ids = this.todos.filter((t) => t.selected).map((t) => t._id);
     this.api.delete('todos', ids).subscribe(
         (res) => {
@@ -60,8 +101,24 @@ export class TodoComponent implements OnInit {
             const todo = this.todos[t];
             if (todo.selected) this.todos.splice(t, 1);
           }
+          this.enableDeleteMulti = false;
         }
       );
+  }
+
+  /**
+  * [set Filter for elements]
+  */
+  setFilter(filters) {
+    this.enableDeleteMulti = false;
+    this.api.getAll('todos', filters).subscribe((todos) => this.todos = todos);
+  }
+  /**
+  * [reset the filter]
+  */
+  resetFilter() {
+    this.enableDeleteMulti = false;
+    this.api.getAll('todos').subscribe((todos) => this.todos = todos);
   }
 
 }
